[light-client] add Schnorr keys to the HotShot realm (#2762)
* adding schnorr keys to nodetype

* keys!!

* fmt

* migration

* state_cert storage api & test

* sqlite cargo.lock

* fix sql & new test

+++ contracts/rust/adapter/src/light_client.rs
@@ -54,8 +54,8 @@ impl From<contract_bindings_ethers::light_client::LightClientState> for ParsedLi
 impl<F: PrimeField> From<ParsedLightClientState> for GenericLightClientState<F> {
     fn from(v: ParsedLightClientState) -> Self {
         Self {
-            view_number: v.view_num as usize,
-            block_height: v.block_height as usize,
+            view_number: v.view_num,
+            block_height: v.block_height,
             block_comm_root: u256_to_field(v.block_comm_root),
         }
     }
@@ -64,8 +64,8 @@ impl<F: PrimeField> From<ParsedLightClientState> for GenericLightClientState<F> 
 impl<F: PrimeField> From<GenericLightClientState<F>> for ParsedLightClientState {
     fn from(v: GenericLightClientState<F>) -> Self {
         Self {
-            view_num: v.view_number as u64,
-            block_height: v.block_height as u64,
+            view_num: v.view_number,
+            block_height: v.block_height,
             block_comm_root: field_to_u256(v.block_comm_root),
         }
     }
